pyproject.toml
[tool.poetry]
name = "feedlocal"
version = "0.0.1"
description = "Backend part of project FeedLocal"
authors = ["Suprun Sergey <serhii@peppernode.net>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.1"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
django-tenants = "^3.5.0"
psycopg2-binary = "^2.9.6"
drf-spectacular = "^0.26.2"
python-dotenv = "^1.0.0"
uvicorn = "^0.22.0"
django-filter = "^23.2"
django-cors-headers = "^4.0.0"
boto3 = "^1.26.147"
django-storages = "^1.13.2"
pillow = "^9.5.0"
django-resized = "^1.0.2"
flake8 = "^6.0.0"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
pyyaml = "^6.0"

[tool.black]
line-length = 120

[build-system]
build-backend = "poetry.core.masonry.api"